Nathanael/mpptcleanup (#42)
* Remove span from data

From previous experimentation with layouts but was scrapped. Only 1 channel was changed to match the rest.

* Fix alive css typo and remove extra span around each alive green circle

Alive typo didn't change data but merely the colour of the font, and extra span looked as though it wasn't doing anything and was tested for changes without. No difference and therefore was removed.

* Fix mppt spacing to tabs

The original spaces caused offsets for other computers