Create data service initializer service
The data services were not being initialized on startup.
They were initialized when the component using them was constructed.
This was causing issues as the faults page was not being created by the side
panels, like the other ones. The result was that if you loaded the program
on any other page (eg. MPPT), switching to the faults page would not display
any new data.

We need to implement a initializer class, and add that as a provider
with APP_INITIALIZER to run on startup.

References:
https://stackoverflow.com/a/44731279
https://devblog.dymel.pl/2017/10/17/angular-preload/