Fix secondsdifference returning a bigint (#137)
By default, `lastLapTimestamp` was being set to 0. However, this created an issue for the first lap being put in the database. Since `lastLapTimestamp` was 0, `secondsDifference = timestamp - lastLapTimestamp = timestamp - 0 = timestamp`. However, timestamp is a bigint (about 40 bits) and secondsDifference is only an int (max 32 bits). This caused an error to be thrown when trying to insert the row into the database