Redo distance calculation & add unit tests for it
The original assumption that an odometer can't decrease in value was wrong :( , so it caused massive discrepancies. We were also not rounding the values from meters to kilometers.

I also added some chai unit tests, that can be run by installing mocha and running `./node_modules/mocha/bin/mocha` in the server directory.

I'll do a future PR for documentation & getting travis to run it the unit tests